// backend/src/data/interfaces/month.service.protocol.ts
import { IExcelModel } from "../mapper/excel.mapper"

export interface Months {
  jan: IExcelModel[]
  fev: IExcelModel[]
  mar: IExcelModel[]
  abr: IExcelModel[]
  mai: IExcelModel[]
  jun: IExcelModel[]
  jul: IExcelModel[]
  ago: IExcelModel[]
  set: IExcelModel[]
  out: IExcelModel[]
  nov: IExcelModel[]
  dez: IExcelModel[]
}